remove 'deploy staging' step
we'll provision the staging environment with the latest version of the
app and shut it down immediately after. as staging is not supposed to
run 24/7 for the moment, we'll stop deploying to it as part of our
regular pipeline.

whenever there'll be a 24/7 available staging environment we can bring
this step back to live.