refactor(affinities): move affinites from player into behavior
kinda bigger refactor. affinities shouldn't be part of a player, instead they should be part of the behaviour,
serving as storage. game shouldn't handle affinityevents, instead they should be passed to behavior,
since actions are based on behavior.
Circular dependency between game and player is still existing in the codebase and needs refactoring